Update package.yml
Hey @p3lim, am I doing this right? It just comes back with "Failed" -- I had set this up before, and it was just copy and paste and set up my variables in the ToC and Secrets... I don't use WoW Interface or Wago... do I need to provide a GitHub token? I don't see why it would have to write to GitHub... I have been having trouble getting this to work on any add-ons, but this one is probably a bit more complex in that it publishes to Retail, Wrath, and Classic Era... I set up the ToC files for each. What am I missing?